+++ lib/cucumber/formatter/http_io.rb
@@ -2,16 +2,15 @@ require 'net/http'
 
 module Cucumber
   module Formatter
-    # Quacks like an IO and executes a HTTP request with the data as body on flush
     class HTTPIO
-      # attr_reader :req
-
       class << self
+        # Returns an IO that will write to a HTTP request's body
         def open(url, https_verify_mode = nil)
           @https_verify_mode = https_verify_mode
-          uri, headers = build_uri_and_headers(url)
+          uri, method, headers = build_uri_and_headers(url)
 
-          @req = Net::HTTP::Post.new(uri)
+          method_class_name = "#{method[0].upcase}#{method[1..-1].downcase}"
+          @req = Net::HTTP.const_get(method_class_name).new(uri)
           headers.each do |header, value|
             @req[header] = value
           end
@@ -36,24 +35,24 @@ module Cucumber
           uri = URI(url)
           query_pairs = uri.query ? URI.decode_www_form(uri.query) : []
 
-          # Build headers from query parameters prefixed with http-
+          # Build headers from query parameters prefixed with http- and extract HTTP method
           http_query_pairs = query_pairs.select { |pair| pair[0] =~ /^http-/ }
-          http_query_pairs_wthout_prefix = http_query_pairs.map do |pair|
-            [
-              pair[0][5..-1].downcase, # remove http- prefix
-              pair[1]
-            ]
-          end
+          http_query_hash_without_prefix = Hash[http_query_pairs.map do |pair|
+                                                  [
+                                                    pair[0][5..-1].downcase, # remove http- prefix
+                                                    pair[1]
+                                                  ]
+                                                end]
+          method = http_query_hash_without_prefix.delete('method') || 'POST'
           headers = {
-            'content-type' => 'application/json',
             'transfer-encoding' => 'chunked'
-          }.merge(Hash[http_query_pairs_wthout_prefix])
+          }.merge(http_query_hash_without_prefix)
 
           # Update the query with the http-* parameters removed
           remaining_query_pairs = query_pairs - http_query_pairs
           new_query_hash = Hash[remaining_query_pairs]
           uri.query = URI.encode_www_form(new_query_hash) unless new_query_hash.empty?
-          [uri, headers]
+          [uri, method, headers]
         end
       end
     end