+++ lib/cucumber/formatter/message_builder.rb
@@ -47,6 +47,7 @@ module Cucumber
         }
 
         if media_type.start_with?('text/')
+          attachment_data[:content_encoding] = Cucumber::Messages::Attachment::ContentEncoding::IDENTITY
           attachment_data[:body] = src
         else
           body = src.respond_to?(:read) ? src.read : src
@@ -125,7 +126,7 @@ module Cucumber
 
       def step_match_arguments(step)
         @step_definitions_by_test_step.step_match_arguments(step).map do |argument|
-          Cucumber::Messages::StepMatchArgument.new(
+          Cucumber::Messages::TestCase::TestStep::StepMatchArgumentsList::StepMatchArgument.new(
             group: argument_group_to_message(argument.group),
             parameter_type_name: argument.parameter_type.name
           )
@@ -133,7 +134,7 @@ module Cucumber
       end
 
       def argument_group_to_message(group)
-        Cucumber::Messages::StepMatchArgument::Group.new(
+        Cucumber::Messages::TestCase::TestStep::StepMatchArgumentsList::StepMatchArgument::Group.new(
           start: group.start,
           value: group.value,
           children: group.children.map { |child| argument_group_to_message(child) }
@@ -188,7 +189,7 @@ module Cucumber
 
         result_message = result.to_message
         if result.failed? || result.pending?
-          result_message = Cucumber::Messages::TestStepResult.new(
+          result_message = Cucumber::Messages::TestStepFinished::TestStepResult.new(
             status: result_message.status,
             duration: result_message.duration,
             message: create_error_message(result)