+++ CHANGELOG.md
@@ -9,12 +9,15 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CONTRIBUTING.md) for more info on how to contribute to Cucumber.
 
 ----
-
 ## [In GIT](https://github.com/cucumber/cucumber-ruby/compare/v4.0.0.rc.4...master)
 
 ### Added
 
-* N/A
+* Accept `--out URL` to POST results to a web server
+  If a URL is used as output, the output will be sent with a POST request. 
+  This can be overridden by specifying e.g. `http-method=PUT` as a query parameter. 
+  Other `http-` prefixed query parameters will be converted to request headers 
+  (with the `http-` prefix stripped off).
 
 ### Changed
 
@@ -27,7 +30,6 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 ### Improved
 
 * N/A
-
 
 ## [4.0.0.rc.4](https://github.com/cucumber/cucumber-ruby/compare/v4.0.0.rc.3...4.0.0.rc.4)