+++ CHANGELOG.md
@@ -22,6 +22,12 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
    [#1402](https://github.com/cucumber/cucumber-ruby/pull/1402)
    [karamosky](https://github.com/karamosky))
 
+* Updated monorepo libraries:
+  - cucumber-gherkin ~> 12
+  - cucumber-html-formatter ~> 5
+
+* Use `body` field of attachments
+
 ### Removed
 
 * N/A