+++ CHANGELOG.md
@@ -17,14 +17,15 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Changed
 
-* Code snippet for an undefined step with a Doc String will ouput `doc_string` instead of `string` in block params 
+* Code snippet for an undefined step with a Doc String will ouput `doc_string` instead of `string` in block params
   ([#1401](https://github.com/cucumber/cucumber-ruby/issues/1401)
    [#1402](https://github.com/cucumber/cucumber-ruby/pull/1402)
    [karamosky](https://github.com/karamosky))
 
 * Updated monorepo libraries:
-  - cucumber-gherkin ~> 12
-  - cucumber-html-formatter ~> 5
+  - cucumber-gherkin ~> 13
+  - cucumber-html-formatter ~> 6
+  - cucumber-cucumber-expressions ~> 10
 
 * Use `body` field of attachments